refactor(chatDialog): move list of messages into its own file
